// src/app/modules/monitoring/grading/model/step.ts
import { StepProgress } from "./step-progress";
import { Standard } from "./strandard";
import { SubmissionFormat } from "./submission-format";

export interface Step {
    id?: number;
    parentId?: number;
    order: number;
    code: string;
    name: string;
    submissionFormat: SubmissionFormat;
    standards?: Standard[];
    progress?: StepProgress;
}